OHRM5X-519: Code clean ups, AssignLeaveAPITest.yml file renamed as AssignLeaveAPITestCases.yml and AssignLeaveAPITest.php file removed
